feat: :recycle: Introduce abstact-class `Converter`
