+++ pyproject.toml
@@ -104,7 +104,8 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
 spectrafit = "spectrafit.spectrafit:command_line_runner"
-spectrafit-converter = "spectrafit.plugins.converter:command_line_runner"
+spectrafit-input-converter = "spectrafit.plugins.input_converter:command_line_runner"
+spectrafit-data-converter = "spectrafit.plugins.data_converter:command_line_runner"
 spectrafit-jupyter = "spectrafit.app.app:jupyter"
 
 [tool.poetry.extras]
