+++ spectrafit/plugins/data_converter.py
@@ -1,0 +1,1 @@
+"""Transform the input data to a CSV file."""
